Changing resolveImpl to be closer to the Java semantics.
Because new DirectoryInfo objects are created relative to
the working directory, the Exists check would examine a
different file than would be returned if it did not exist.

For example:
working directory is c:\tests\bin

resolveImpl("c:\test\bin\.git", "objects");

would check if "c:\test\bin\objects" existed and return it
if it did. And if it did not exist it would return
"c:\test\bin\.git\objects". The variance in the return values
was probably unintended.