Fixing test020_createBlobTag and related tests
ObjectWriter.WriteCommit was using a StreamWriter and a Binary Writer
on the same MemoryStream at the same time. This had a couple of
problems.

StreamWriter keeps position information about where to write in to the
stream. The two writers were overwriting each other. As far as I can
tell, the Writers expect to have exclusive access to their underlying
stream, although I couldn't find a specific statement about this in the
documentation.

The second problem came after removing the StreamWriter and trying
to use only the BinaryWriter. BinaryWriter.Write(string) prepends a
length number to any string it writes. (This is pascal style strings?)
To avoid this I'm calling .ToCharArray() before passing strings so that
the BinaryWriter.Write(char[]) method is selected, which has the simple
behaviour.

The .ToCharArray() calls aren't a good long term solution as they will
be very easy to forget. I'm wondering if a GitFormatWriter or similar
should be introduced. After the porting is complete?