Fix for equals operator with null comparisons
When comparing a null AbstractObject with null using the
equals operator a NullReferenceException would be thrown;
this was due to Equals being called on the (null) instance.
Now Object.Equals is used to compare the two, which
internally does a null check prior to calling the virtual
Equals method on the instance.