Fix for extensionless files always showing as modified in the index.
As per http://bit.ly/1cDvWA, this was because the FileSystemInfo was
unable to tell whether an extensionless file was a file or a directory
by name alone. This caused the file name to be appended to itself, and
thus never appeared to exist. An IsFile check has been introduced to
first verify whether the path is a file regardless of extension.